backend/requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
anykeystore==0.2
apex==0.9.10.dev0
asgiref==3.6.0
astunparse==1.6.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
blobfile==2.1.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.1.0
cryptacular==1.6.2
cryptography==42.0.2
cycler==0.11.0
Cython==3.0.11
daphne==4.0.0
dataclasses-json==0.5.14
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
Django==4.1.6
django-bmemcached==0.3.0
django-heroku==0.3.1
django-redis-cache==1.8.0
django-rosetta==0.9.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.6
fastapi-mail==1.4.2
filelock==3.12.2
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.0
fpdf2==2.7.7
frozenlist==1.4.1
fsspec==2024.12.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.27.1
hupper==1.12.1
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.4
langchain==0.0.124
libclang==16.0.6
lxml==4.9.4
Mako==1.3.8
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
passlib==1.7.4
PasteDeploy==3.1.0
pbkdf2==1.3
pdfkit==1.0.0
pdfminer==20191125
pdfminer.six==20231228
Pillow==10.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
polib==1.2.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pylatexenc==2.10
pymemcache==4.0.0
pyOpenSSL==24.0.0
pyparsing==3.0.9
pyramid==2.0.2
pyramid-mailer==0.15.1
python-binary-memcached==0.31.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
regex==2023.12.25
repoze.sendmail==4.4.1
requests==2.23.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.5.2
scipy==1.12.0
service-identity==24.1.0
six==1.10.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.41.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.1.1
tokenizers==0.21.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transaction==4.0
transformers==4.48.0
translationstring==1.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uhashring==2.3
uritemplate==4.1.1
urllib3==1.25.11
uvicorn==0.34.0
velruse==1.1.1
venusian==3.1.0
virtualenv==20.25.1
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.15.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
yarl==1.9.4
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
